deployment/morph: split off release.nix
Restructure the deployment code to provide a release.nix that can be
used in CI to build all machines that are going to be deployed.

- The new `machines.nix` defines which machines actually get deployed
- morph uses `network.nix` which imports `machines.nix`
- The new `release.nix` also imports `machines.nix` but with some
  changes so it can be passed to `nix-build`

So `machines.nix` defines the configurations to build/deploy in both
cases but for release.nix the `mkMachine` invocation is changed to
provide a fake `deployment` NixOS option because it would otherwise
complain about an unknown option.