Eliminate ReaderT in the CEK machine
This is another one of Ed's suggestions. The use of `ImplicitParams` is
inessential: the real point here is using function args instead of
`ReaderT`.

I'm not sure this is quite the way to go, in particular I killed
`SpendBudget` in favour of just passing a spending function, since it
only seemed to be used in one place by `applyTypeSchemed`.

Seems to give a ~5% performance improvement (benchmarks in a comment).