SCP-2289: Try making a special "for evaluation" term type
Another one of Ed's suggestions. I got rid of annotations and replaced
names with `Unique`s directly, which we can even unpack into the term.

This seems to have better Core: indeed, it gets rid of some wrapping and
unwrapping when looking up or inserting variable mappings.

But the result is overall much worse.